models.py transformed into package with files
