+++ backend/api/clients/fetch.py
@@ -3,7 +3,7 @@ from django.http import HttpRequest
 from django.shortcuts import render, redirect
 from django.views.decorators.http import require_http_methods
 
-from backend.models_db.client import Client
+from backend.models.client import Client
 
 
 @require_http_methods(["GET"])
