+++ backend/models/feature_flags.py
@@ -1,0 +1,14 @@
+from django.db import models
+
+
+class FeatureFlags(models.Model):
+    name = models.CharField(max_length=100)
+    value = models.BooleanField(default=False)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        verbose_name = "Feature Flag"
+        verbose_name_plural = "Feature Flags"
+
+    def __str__(self):
+        return self.name
