+++ backend/models/user.py
@@ -1,0 +1,161 @@
+from django.db import models
+from uuid import uuid4
+from django.contrib.auth.models import UserManager, AbstractUser, AnonymousUser
+from django.db.models import Count
+from settings import settings
+from django.contrib.auth.hashers import make_password
+from backend.models.utils import RandomCode, add_3hrs_from_now
+
+
+class CustomUserManager(UserManager):
+    def get_queryset(self):
+        return (
+            super()
+            .get_queryset()
+            .select_related("user_profile", "logged_in_as_team")
+            .annotate(notification_count=((Count("user_notifications"))))
+        )
+
+
+class User(AbstractUser):
+    objects = CustomUserManager()
+
+    logged_in_as_team = models.ForeignKey("Team", on_delete=models.SET_NULL, null=True, blank=True)
+    awaiting_email_verification = models.BooleanField(default=True)
+
+    class Role(models.TextChoices):
+        #        NAME     DJANGO ADMIN NAME
+        DEV = "DEV", "Developer"
+        STAFF = "STAFF", "Staff"
+        USER = "USER", "User"
+        TESTER = "TESTER", "Tester"
+
+    role = models.CharField(max_length=10, choices=Role.choices, default=Role.USER)
+
+
+class CustomUserMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        # Replace request.user with CustomUser instance if authenticated
+        if request.user.is_authenticated:
+            request.user = User.objects.get(pk=request.user.pk)
+        else:
+            # If user is not authenticated, set request.user to AnonymousUser
+            request.user = AnonymousUser()
+
+        response = self.get_response(request)
+        return response
+
+
+class VerificationCodes(models.Model):
+    class ServiceTypes(models.TextChoices):
+        CREATE_ACCOUNT = "create_account", "Create Account"
+        RESET_PASSWORD = "reset_password", "Reset Password"
+
+    uuid = models.UUIDField(default=uuid4, editable=False, unique=True)  # This is the public identifier
+    token = models.TextField(default=RandomCode, editable=False)  # This is the private token (should be hashed)
+
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    created = models.DateTimeField(auto_now_add=True)
+    expiry = models.DateTimeField(default=add_3hrs_from_now)
+    service = models.CharField(max_length=14, choices=ServiceTypes.choices)
+
+    def __str__(self):
+        return self.user.username
+
+    def is_expired(self):
+        if timezone.now() > self.expiry:
+            self.delete()
+            return True
+        return False
+
+    def hash_token(self):
+        self.token = make_password(self.token)
+        self.save()
+        return True
+
+    class Meta:
+        verbose_name = "Verification Code"
+        verbose_name_plural = "Verification Codes"
+
+
+class UserSettings(models.Model):
+    CURRENCIES = {
+        "GBP": {"name": "British Pound Sterling", "symbol": "£"},
+        "EUR": {"name": "Euro", "symbol": "€"},
+        "USD": {"name": "United States Dollar", "symbol": "$"},
+        "JPY": {"name": "Japanese Yen", "symbol": "¥"},
+        "INR": {"name": "Indian Rupee", "symbol": "₹"},
+        "AUD": {"name": "Australian Dollar", "symbol": "$"},
+        "CAD": {"name": "Canadian Dollar", "symbol": "$"},
+    }
+    user = models.OneToOneField(User, on_delete=models.CASCADE, related_name="user_profile")
+    dark_mode = models.BooleanField(default=True)
+    currency = models.CharField(
+        max_length=3,
+        default="GBP",
+        choices=[(code, info["name"]) for code, info in CURRENCIES.items()],
+    )
+    profile_picture = models.ImageField(
+        upload_to="profile_pictures/",
+        storage=settings.CustomPublicMediaStorage(),
+        blank=True,
+        null=True,
+    )
+
+    @property
+    def profile_picture_url(self):
+        if self.profile_picture and hasattr(self.profile_picture, "url"):
+            return self.profile_picture.url
+        return ""
+
+    def get_currency_symbol(self):
+        return self.CURRENCIES.get(self.currency, {}).get("symbol", "$")
+
+    def __str__(self):
+        return self.user.username
+
+    class Meta:
+        verbose_name = "User Settings"
+        verbose_name_plural = "User Settings"
+
+
+class Team(models.Model):
+    name = models.CharField(max_length=100, unique=True)
+    leader = models.ForeignKey(User, on_delete=models.CASCADE, related_name="teams_leader_of")
+    members = models.ManyToManyField(User, related_name="teams_joined")
+
+
+class TeamInvitation(models.Model):
+    code = models.CharField(max_length=10)
+    team = models.ForeignKey(Team, on_delete=models.CASCADE, related_name="team_invitations")
+    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="team_invitations")
+    invited_by = models.ForeignKey(User, on_delete=models.CASCADE)
+    expires = models.DateTimeField(null=True, blank=True)
+    active = models.BooleanField(default=True)
+
+    def is_active(self):
+        if not self.active:
+            return False
+        if timezone.now() > self.expires:
+            self.active = False
+            self.save()
+            return False
+        return True
+
+    def set_expires(self):
+        self.expires = timezone.now() + timezone.timedelta(days=7)
+
+    def save(self, *args, **kwargs):
+        self.set_expires()
+        self.code = RandomCode(10)
+        super().save()
+
+    def __str__(self):
+        return self.team.name
+
+    class Meta:
+        verbose_name = "Team Invitation"
+        verbose_name_plural = "Team Invitations"
