+++ backend/signals.py
@@ -13,7 +13,7 @@ from django.urls import reverse
 
 import settings.settings
 from backend.models import UserSettings, User, FeatureFlags, VerificationCodes, QuotaLimit
-from backend.models_db.receipt import Receipt
+from backend.models.receipt import Receipt
 
 from settings.helpers import ARE_EMAILS_ENABLED, send_email
