+++ backend/views/core/auth/login.py
@@ -14,7 +14,7 @@ from django_ratelimit.decorators import ratelimit
 
 from backend.decorators import *
 from backend.models import User, VerificationCodes
-from backend.models_db.logger import LoginLog, AuditLog
+from backend.models.logger import LoginLog, AuditLog
 from backend.views.core.auth.verify import create_magic_link
 from settings.helpers import send_email, ARE_EMAILS_ENABLED
